Refactor to use bringup launch script with gazebo launch script
The gazebo launch script now only takes care of launching gazebo and the bringup script handles launching nodes that are common between simulation and the real robot and those specific to a particular mode.  This approach removes duplication.

Gazebo can be started from the bringup script or it can be launched independently from the bringup script.  The later case allows gazebo to be run on the dev computer with the other nodes run on the actual robot computer to better test against the limitations of the robot computer.